% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompSize.R
\name{CompSize}
\alias{CompSize}
\title{Calculates CompetitorSize vector for a single date, given similarity and size}
\usage{
CompSize(sim.mat, fund.size)
}
\arguments{
\item{sim.mat}{Matrix of similarity scores.}

\item{fund.size}{Mector of fund sizes.}
}
\value{
Matrix of competitor sizes.
}
\description{
Calculates CompetitorSize vector for a single date, given similarity and size
}
\examples{
w <- cbind(c(0.75, 0.25), c(0.25, 0.75))
colnames(w) <- c("fund1", "fund2")
sim.mat <- CosSim(w)
fund.size <- matrix(c(0.1, 0.2), 2, 1)
rownames(fund.size) <- c("fund1", "fund2")
CompSize(sim.mat, fund.size)
}
