% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitByDate.R
\name{SplitByDate}
\alias{SplitByDate}
\title{Split Weights and Sizes by Date}
\usage{
SplitByDate(pw, size, fund.id = "wficn", date.id = "date",
  stock.id = "permno", w.var = "w", size.var = "fund.size")
}
\arguments{
\item{pw}{data.table (or data.frame) of portfolio weights (or other measure based on which fund similarity should be calculated). Must be a long-form dataset, with columns specifying fund, stock, date, and the  corresponding portfolio weight.}

\item{size}{data.table (or data.frame) of fund sizes. Must be a long-form dataset, with columns corresponding to fund, date, and the size measure.}

\item{fund.id}{Name of the variable (supplied as a character string) containing fund identifiers. Defaults to "wficn".}

\item{date.id}{Name of the variable (supplied as a character string) conteining time identifiers. Defaults to "date".}

\item{stock.id}{Name of the variable (supplied as a character string) containing security identifiers. Defaults to "permno".}

\item{w.var}{Name of the variable (supplied as a character string) based on which fund similarity is to be calculated. Defaults to "w". In Jakab (2018), w.var is market-adjusted portfolio weights.}

\item{fund.size}{Name of the variable (supplied as a character string) based on which competitor size is to be calculated. In Jakab (2018), fund.size is FundSize, i.e. total net assets normalized by the market capitalization of all common equity in CRSP.}
}
\value{
A nested list. Each top level element corresponds to a date (reflected in the list names), and contains three sub-elements. (i) is an N x K matrix of portfolio weights (or measures used for calculating fund similarity); row names contain N stock ids, column names contain K fund ids. (ii) is a K x 1 matrix of fund sizes, with row names containing fund ids. (iii) is the date, in string format.
}
\description{
Split Weights and Sizes by Date
}
\examples{
pw <- data.table(
  wficn  = paste0("fund", rep(1:2, each = 50)),
  date   = paste0("date", rep(rep(1:2, each = 25), 2)),
  permno = paste0("stock", rep(1:25, 4)),
  w      = runif(100))
pw[, w := w / sum(w), by = .(wficn, date)]
pw
size <- data.table(
  wficn = paste0("fund", rep(1:2, each = 2)),
  date  = paste0("date", rep(1:2, 2)),
  fund.size = c(.1, .11, .2, .25))
size
SplitByDate(pw, size)
}
